refactor FixedPool
